requirements.txt
django==4.2
djangorestframework==3.14
djangorestframework-simplejwt==5.3
psycopg2-binary==2.9
python-dotenv==1.0